backend/pyproject.toml
[project]
name = "project-astralbound"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.0",
    "psycopg[binary]>=3.3.3",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.6",
    "sqlmodel>=0.0.34",
    "uvicorn[standard]>=0.40.0",
]

# TODO: Add more ruff settings
# TODO: Add support for docstring
